Upload logo

+++ app/Controllers/Http/CartsController.ts
@@ -1,6 +1,7 @@
 import { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
 import { schema } from '@ioc:Adonis/Core/Validator'
 import Cart from 'App/Models/Cart'
+import Course from 'App/Models/Course'
 import Product from 'App/Models/Product'
 
 export default class CartsController {
@@ -8,44 +9,81 @@ export default class CartsController {
     const carts = await Cart.query()
       .where('user_id', auth.use('userApi').user?.id!)
       .preload('product')
+      .preload('course')
 
     return carts.map((cart) => {
       const serialize = cart.serialize()
+      let item: any
+
+      if (serialize.type === 'product') {
+        item = serialize.product
+      } else if (serialize.type === 'course') {
+        item = serialize.course
+      }
       return {
-        id: serialize.product?.id,
+        id: item.id,
         qty: serialize.qty,
-        name: serialize.product?.title,
-        price: serialize.product?.price,
-        image: serialize.product?.cover,
+        name: item.title,
+        price: item.price,
+        image: item.cover,
+        type: serialize.type,
       }
     })
   }
 
-  public async increase({ auth, request }: HttpContextContract) {
-    const { qty, product_id } = await request.validate({
+  public async increase({ auth, request, response }: HttpContextContract) {
+    const {
+      qty = 1,
+      product_id,
+      course_id,
+    } = await request.validate({
       schema: schema.create({
-        qty: schema.number(),
-        product_id: schema.number(),
+        qty: schema.number.optional(),
+        product_id: schema.number.optional(),
+        course_id: schema.number.optional(),
       }),
     })
 
-    const product = await Product.findByOrFail('id', product_id)
-    let cart = await Cart.query()
-      .where('product_id', product.id)
-      .andWhere('user_id', auth.use('userApi').user?.id!)
-      .first()
-    if (!cart) {
-      cart = new Cart()
-      cart.userId = auth.use('userApi').user?.id!
-      cart.productId = product.id
-      cart.qty = qty || 1
-    } else {
-      cart.qty = cart.qty + qty
+    let item: Product | Course | null = null
+    let cart: Cart | null = null
+
+    if (product_id) {
+      item = await Product.findOrFail(product_id)
+      cart = await Cart.query()
+        .where('user_id', auth.use('userApi').user?.id!)
+        .andWhere('product_id', item.id)
+        .first()
+
+      if (!cart) {
+        cart = new Cart()
+        cart.userId = auth.use('userApi').user?.id!
+        cart.productId = item.id
+        cart.qty = qty || 1
+      } else {
+        cart.qty = cart.qty + qty
+      }
+    } else if (course_id) {
+      item = await Course.findOrFail(course_id)
+      cart = await Cart.query()
+        .where('user_id', auth.use('userApi').user?.id!)
+        .andWhere('course_id', item.id)
+        .first()
+
+      if (!cart) {
+        cart = new Cart()
+        cart.userId = auth.use('userApi').user?.id!
+        cart.courseId = item.id
+        cart.qty = 1
+      } else {
+        cart.qty = cart.qty + qty
+      }
     }
 
-    await cart.save()
+    if (!item) return response.badRequest()
 
-    return cart.serialize()
+    await cart!.save()
+
+    return cart!.serialize()
   }
 
   public async decrease({ auth, request }: HttpContextContract) {
@@ -55,8 +93,6 @@ export default class CartsController {
         product_id: schema.number(),
       }),
     })
-
-    console.log(auth.use('userApi').user?.id!, product_id)
 
     const cart = await Cart.query()
       .where('product_id', product_id)
@@ -73,23 +109,34 @@ export default class CartsController {
     return cart.serialize()
   }
 
-  public async reset({ auth, request }: HttpContextContract) {
-    const { qty, product_id } = await request.validate({
+  public async reset({ auth, request, response }: HttpContextContract) {
+    const {
+      qty = 0,
+      product_id,
+      course_id,
+    } = await request.validate({
       schema: schema.create({
-        qty: schema.number(),
-        product_id: schema.number(),
+        qty: schema.number.optional(),
+        product_id: schema.number.optional(),
+        course_id: schema.number.optional(),
       }),
     })
 
-    const cart = await Cart.query()
-      .where('product_id', product_id)
-      .andWhere('user_id', auth.use('userApi').user?.id!)
-      .firstOrFail()
+    let cart: Cart | null = null
+    let cartQuery = Cart.query().where('user_id', auth.use('userApi').user?.id!)
+
+    if (product_id) {
+      cart = await cartQuery.andWhere('product_id', product_id).firstOrFail()
+    } else if (course_id) {
+      cart = await cartQuery.andWhere('course_id', course_id).firstOrFail()
+    }
+
+    if (!cart) return response.badRequest()
 
     if (qty <= 0) {
       await cart.delete()
     } else {
-      cart.qty = qty
+      cart.qty = product_id ? qty : 1
       await cart.save()
     }
